disable fault domain script by default
fault domain is now a required option in EE DC/OS.
This change disables the fault domains by default, however it is up to the user
to override it from a specific test. If we ever need to test a fault domain
in EE, the user might override this change from extra config.
https://jira.mesosphere.com/browse/DCOS-19702