Add master host port map for Docker
Add an option to expose host ports for one of the master node. This is
useful on OSX on which the container IP is not directly accessible from
the host. By exposing the host ports, the user can reach the services on
the master node using the mapped host ports. The host port map will be
applied to one master only if there are multiple master nodes.