Block device support using loopback devices.
This makes it possible to provide an unformatted block device to a cluster
that is using the Docker backend.
Support for this is added to the API (via `DockerLoopbackVolume`) as well as
the `dcos-docker` CLI (via `create-loopback-sidecar`/`destroy-loopback-sidecar`
commands). It creates a sidecar container containing an unformatted file that is
mounted to all nodes as a loopback device.